# app/routes/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Designed to be consumed by the Integrated Pipeline Reports (IPR) Client, the endpoints are\navailable for authorized users to consume.",
    "version": "1.0.0",
    "title": "Integrated Pipeline Reports (IPR)",
    "contact": {
      "name": "IPR Team",
      "email": "ipr@bcgsc.ca"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "paths": {
...
    "/reports": {
      "get": {
        "summary": "Get Genomic Reports",
        "description": "Retrieve all genomic reports that you have acess to",
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "project",
            "in": "query",
            "required": false,
            "description": "Seach for all reports under a chosen project name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "searchText",
            "in": "query",
            "required": false,
            "description": "String to be searched in the reports field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "states",
            "in": "query",
            "required": false,
            "description": "Filter by current state of the report",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "description": "Filter the reports by user role",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "paginated",
            "in": "query",
            "required": false,
            "description": "'true', if you want to split the report, allowing you to choose the 'limit' and 'offset",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "The number of reports to retrieve",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "The start of the set of reports to retrieve",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "description": "Column(s) for the reports to be sorted on",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object with the number of reports and an array of genomic reports",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "reports": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/analysis_reportAssociations"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "Insufficient access to get reports"
          }
        }
      },
      "post": {
        "summary": "Add Genomic Report",
        "description": "Upload a genomic report and all of the report sections",
        "tags": [
          "Reports"
        ],
        "parameters": [
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/analysis_reportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns an object containing a message and the ident of the new report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "ident": {
                      "$ref": "#/components/schemas/ident"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request syntax"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/reports/{report}": {
      "get": {
        "summary": "Get Genomic Report",
        "description": "Retrieve a specific genomic report",
        "parameters": [
          {
            "$ref": "#/components/parameters/report"
          }
        ],
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the details of the specified genomic report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/analysis_reportAssociations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Report not found"
          }
        }
      },
      "put": {
        "summary": "Update Genomic Report",
        "description": "Update the details of a genomic report",
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/report"
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/analysis_reportUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the details of the updated genomic report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/analysis_reportAssociations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Report not found"
          }
        }
      },
      "delete": {
        "summary": "Delete Genomic Report",
        "description": "Removes the specified genomic report",
        "parameters": [
          {
            "$ref": "#/components/parameters/report"
          }
        ],
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Report not found"
          }
        }
      }
    },
...
  },
...
  "components": {
    "parameters": {
      "report": {
        "name": "report",
        "in": "path",
        "required": true,
        "description": "Analysis report ident",
        "schema": {
          "type": "string",
          "format": "UUIDv4"
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
      "ident": {
        "type": "string",
        "format": "UUIDv4",
        "example": "8cd64045-7e68-4455-bcc2-e750c4372974",
        "description": "Randomly generated public unique ID"
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication information is missing or invalid",
        "headers": {
          "WWW_Authenticate": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}